backend/pyproject.toml
[project]
name = "quantsage"
version = "0.1.0"
description = "AI Crypto Trader — Risk-First, Explainable, Regime-Aware"
requires-python = ">=3.12"
readme = "../README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "ccxt>=4.4",
    "pyupbit>=0.2.34",
    "pandas>=2.2",
    "numpy>=2.0",
    "pandas-ta>=0.3.14b0",
    "scipy>=1.14",
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "statsmodels>=0.14",
    "arch>=7.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "apscheduler>=3.10",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "python-telegram-bot>=21",
    "httpx>=0.27",
    "websockets>=13",
    "orjson>=3.10",
    "loguru>=0.7",
    "prometheus-client>=0.21",
    "python-dotenv>=1.0",
    "cryptography>=43",
    "anthropic>=0.40",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.12",
    "PyJWT>=2.9",
    "aiolimiter>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.11",
    "httpx>=0.27",
]
ml = [
    "torch>=2.5",
    "transformers>=4.45",
    "shap>=0.46",
    "optuna>=4.0",
    "vectorbt>=0.26",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "ASYNC"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["quantsage*"]
